+++ semcore/notice/src/NoticeSmart.jsx
@@ -12,12 +12,26 @@ class NoticeSmart extends Component {
 
   render() {
     const SNoticeSmart = Root;
-    const { Children, label, actions, closable, onClose } = this.asProps;
+    const { Children, label, title, actions, closable, onClose } = this.asProps;
+    let textContent = <Children />;
+
+    if (typeof Children.origin === 'string') {
+      textContent =
+        isNode(title) || isNode(label) || closable ? (
+          <Notice.Text>
+            <Children />
+          </Notice.Text>
+        ) : (
+          <Children />
+        );
+    }
+
     return (
-      <SNoticeSmart render={Notice}>
+      <SNoticeSmart render={Notice} __excludeProps={['title']}>
         {isNode(label) && <Notice.Label>{label}</Notice.Label>}
         <Notice.Content>
-          <Children />
+          {isNode(title) && <Notice.Title>{title}</Notice.Title>}
+          {textContent}
           {isNode(actions) && <Notice.Actions>{actions}</Notice.Actions>}
         </Notice.Content>
         {closable && <Notice.Close onClick={onClose} />}