+++ website/docs/.vitepress/sidebarConfig.ts
@@ -366,14 +366,14 @@ export const sideBarConfig: SidebarConfig = [
             text: 'Radio',
           },
           {
-            link: '/components/select/select',
-            activeMatch: '/components/select/',
-            text: 'Select / Multiselect',
-          },
-          {
             link: '/components/scroll-area/scroll-area',
             activeMatch: '/components/scroll-area/',
             text: 'ScrollArea',
+          },
+          {
+            link: '/components/select/select',
+            activeMatch: '/components/select/',
+            text: 'Select / Multiselect',
           },
           {
             link: '/components/side-panel/side-panel',
@@ -636,6 +636,11 @@ export const sideBarConfig: SidebarConfig = [
           {
             link: '/filter-group/filter-rules/filter-rules',
             text: 'Filter common rules',
+          },
+          {
+            link: '/filter-group/add-filter/add-filter',
+            activeMatch: '/filter-group/add-filter/',
+            text: 'Add filter',
           },
           {
             link: '/filter-group/advanced-filters/advanced-filters',